Move main logging logic to main-process
